authenticationWrapper class moved to seperate file
