+++ lib/main.dart
@@ -2,8 +2,7 @@ import 'package:events_app/providers/eventProvider.dart';
 import 'package:events_app/providers/societyProvider.dart';
 import 'package:events_app/providers/userProvider.dart';
 import 'package:events_app/auth/authentication_service.dart';
-import 'package:events_app/screens/homePage.dart';
-import 'package:events_app/screens/loginPage.dart';
+import 'package:events_app/widgets/authenticationWrapper.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
@@ -43,16 +42,5 @@ class MyApp extends StatelessWidget {
       ),
       home: AuthenticationWrapper(),
     );
-  }
-}
-
-class AuthenticationWrapper extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    User? _firebaseUser = context.watch<User?>();
-    if (_firebaseUser != null) {
-      return HomePage();
-    }
-    return LoginPage();
   }
 }
